cli: Move command line code into cli package
