+++ cmd/meta/main.go
@@ -20,54 +20,13 @@
 package main
 
 import (
-	"bufio"
-	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"net/http"
 	"os"
-	"os/signal"
-	"strings"
-	"syscall"
 
-	"github.com/docopt/docopt-go"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore/fs"
-	_ "github.com/lib/pq"
-	_ "github.com/mattn/go-sqlite3"
 	"github.com/meta-network/go-meta"
-	"github.com/meta-network/go-meta/cwr"
-	"github.com/meta-network/go-meta/ern"
-	"github.com/meta-network/go-meta/musicbrainz"
-	"github.com/meta-network/go-meta/xml"
+	"github.com/meta-network/go-meta/cli"
 )
-
-var usage = `
-usage: meta import xml <file> [<context>...]
-       meta import xsd <name> <uri> [<file>]
-       meta dump [--format=<format>] <path>
-       meta server [--port=<port>] [--musicbrainz-index=<sqlite3-uri>] [--cwr-index=<sqlite3-uri>]
-       meta musicbrainz convert <postgres-uri>
-       meta musicbrainz index <sqlite3-uri>
-       meta cwr convert <file> <cwr-python-dir>
-       meta cwr index <sqlite3-uri>
-       meta ern convert <files>...
-       meta ern index <sqlite3-uri>
-`[1:]
-
-type Main struct {
-	store  *meta.Store
-	stdin  io.Reader
-	stdout io.Writer
-}
-
-func NewMain(store *meta.Store, stdin io.Reader, stdout io.Writer) *Main {
-	return &Main{store, stdin, stdout}
-}
 
 func main() {
 	log.Root().SetHandler(log.StreamHandler(os.Stderr, log.TerminalFormat(true)))
@@ -77,399 +36,9 @@ func main() {
 		log.Crit("error opening meta store", "err", err)
 	}
 
-	m := NewMain(store, os.Stdin, os.Stdout)
-	if err := m.Run(os.Args[1:]...); err != nil {
+	if err := cli.New(store, os.Stdin, os.Stdout).Run(os.Args[1:]...); err != nil {
 		log.Crit("error running meta command", "err", err)
 	}
-}
-
-func (m *Main) Run(cmdArgs ...string) error {
-	v, _ := docopt.Parse(usage, cmdArgs, true, "0.0.1", false)
-	args := Args(v)
-
-	switch {
-	case args.Bool("import"):
-		return m.RunImport(args)
-	case args.Bool("dump"):
-		return m.RunDump(args)
-	case args.Bool("server"):
-		return m.RunServer(args)
-	case args.Bool("musicbrainz"):
-		return m.RunMusicBrainz(args)
-	case args.Bool("cwr"):
-		return m.RunCwr(args)
-	case args.Bool("ern"):
-		return m.RunERN(args)
-	default:
-		return errors.New("unknown command")
-	}
-}
-
-func (m *Main) RunImport(args Args) error {
-	switch {
-	case args.Bool("xml"):
-		return m.RunImportXML(args)
-	case args.Bool("xsd"):
-		return m.RunImportXMLSchema(args)
-	default:
-		return errors.New("unknown import format")
-	}
-}
-
-func (m *Main) RunImportXML(args Args) error {
-	f, err := os.Open(args.String("<file>"))
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	var context []*cid.Cid
-	if contextArg, ok := args["<context>"]; ok {
-		for _, v := range contextArg.([]string) {
-			cid, err := cid.Decode(v)
-			if err != nil {
-				return fmt.Errorf("invalid CID in --context value %q: %s", v, err)
-			}
-			context = append(context, cid)
-		}
-	}
-
-	obj, err := metaxml.EncodeXML(f, context, m.store.Put)
-	if err != nil {
-		return err
-	}
-
-	log.Info("object created", "cid", obj.Cid())
-
-	return nil
-}
-
-func (m *Main) RunImportXMLSchema(args Args) error {
-	var src io.Reader
-	if path := args.String("<file>"); path != "" {
-		f, err := os.Open(path)
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		src = f
-	} else {
-		uri := args.String("<uri>")
-		res, err := http.Get(uri)
-		if err != nil {
-			return err
-		}
-		defer res.Body.Close()
-		if res.StatusCode != http.StatusOK {
-			return fmt.Errorf("unexpected HTTP status from %s: %s", uri, res.Status)
-		}
-		src = res.Body
-	}
-
-	obj, err := metaxml.EncodeXMLSchema(src, args.String("<name>"), args.String("<uri>"))
-	if err != nil {
-		return err
-	}
-
-	if err := m.store.Put(obj); err != nil {
-		return err
-	}
-
-	log.Info("object created", "cid", obj.Cid())
-
-	return nil
-}
-
-func (m *Main) RunDump(args Args) error {
-	path := strings.Split(args.String("<path>"), "/")
-	cid, err := cid.Decode(path[0])
-	if err != nil {
-		return err
-	}
-	obj, err := m.store.Get(cid)
-	if err != nil {
-		return err
-	}
-	if len(path) == 1 {
-		return json.NewEncoder(m.stdout).Encode(obj)
-	}
-	graph := meta.NewGraph(m.store, obj)
-	v, err := graph.Get(path[1:]...)
-	if err != nil {
-		return err
-	}
-	return json.NewEncoder(m.stdout).Encode(v)
-}
-
-func (m *Main) RunServer(args Args) error {
-	var musicbrainzDB *sql.DB = nil
-	var cwrDB *sql.DB = nil
-	if uri := args.String("--musicbrainz-index"); uri != "" {
-		db, err := sql.Open("sqlite3", uri)
-		if err != nil {
-			return err
-		}
-		defer db.Close()
-		musicbrainzDB = db
-	}
-	if uri := args.String("--cwr-index"); uri != "" {
-		db, err := sql.Open("sqlite3", uri)
-		if err != nil {
-			return err
-		}
-		defer db.Close()
-		cwrDB = db
-	}
-	srv, err := NewServer(m.store, musicbrainzDB, cwrDB)
-	if err != nil {
-		return err
-	}
-	port := args.String("--port")
-	if port == "" {
-		port = "5000"
-	}
-	addr := "0.0.0.0:" + port
-	log.Info("starting META HTTP server", "addr", addr)
-	return http.ListenAndServe(addr, srv)
-}
-
-func (m *Main) RunMusicBrainz(args Args) error {
-	switch {
-	case args.Bool("convert"):
-		return m.RunMusicBrainzConvert(args)
-	case args.Bool("index"):
-		return m.RunMusicBrainzIndex(args)
-	default:
-		return errors.New("unknown musicbrainz command")
-	}
-}
-
-func (m *Main) RunMusicBrainzConvert(args Args) error {
-	db, err := sql.Open("postgres", args.String("<postgres-uri>"))
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	// shutdown gracefully on SIGINT or SIGTERM
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		defer cancel()
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		log.Info("received signal, exiting...")
-	}()
-
-	// run the converter in a goroutine so that we only exit once all CIDs
-	// have been read from the stream
-	stream := make(chan *cid.Cid)
-	errC := make(chan error, 1)
-	go func() {
-		defer close(stream)
-		errC <- musicbrainz.NewConverter(db, m.store).ConvertArtists(ctx, stream)
-	}()
-
-	// output the resulting CIDs to stdout
-	for cid := range stream {
-		fmt.Fprintln(m.stdout, cid.String())
-	}
-
-	return <-errC
-}
-
-func (m *Main) RunMusicBrainzIndex(args Args) error {
-	db, err := sql.Open("sqlite3", args.String("<sqlite3-uri>"))
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	indexer, err := musicbrainz.NewIndexer(db, m.store)
-	if err != nil {
-		return err
-	}
-
-	// stream the CIDs from stdin
-	stream := make(chan *cid.Cid)
-	go func() {
-		defer close(stream)
-		s := bufio.NewScanner(m.stdin)
-		for s.Scan() {
-			cid, err := cid.Parse(s.Text())
-			if err != nil {
-				log.Error("error parsing cid", "value", s.Text(), "err", err)
-				return
-			}
-			stream <- cid
-		}
-	}()
-
-	// shutdown gracefully on SIGINT or SIGTERM
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		defer cancel()
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		log.Info("received signal, exiting...")
-	}()
-
-	return indexer.IndexArtists(ctx, stream)
-}
-
-func (m *Main) RunCwr(args Args) error {
-	switch {
-	case args.Bool("convert"):
-		return m.RunCwrConvert(args)
-	case args.Bool("index"):
-		return m.RunCwrIndex(args)
-	default:
-		return errors.New("unknown cwr command")
-	}
-}
-
-func (m *Main) RunCwrConvert(args Args) error {
-	// shutdown gracefully on SIGINT or SIGTERM
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		defer cancel()
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		log.Info("received signal, exiting...")
-	}()
-	file, err := os.Open(args.String("<file>"))
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	// run the converter in a goroutine so that we only exit once all CIDs
-	// have been read from the stream
-	stream := make(chan *cid.Cid)
-	errC := make(chan error, 1)
-	go func() {
-		defer close(stream)
-		errC <- cwr.NewConverter(m.store).ConvertRegisteredWork(ctx, stream, file, args.String("<cwr-python-dir>"))
-	}()
-
-	// output the resulting CIDs to stdout
-	for cid := range stream {
-		fmt.Fprintln(m.stdout, cid.String())
-	}
-
-	return <-errC
-}
-
-func (m *Main) RunCwrIndex(args Args) error {
-
-	db, err := sql.Open("sqlite3", args.String("<sqlite3-uri>"))
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	indexer, err := cwr.NewIndexer(db, m.store)
-	if err != nil {
-		return err
-	}
-
-	// stream the CIDs from stdin
-	stream := make(chan *cid.Cid)
-	go func() {
-		defer close(stream)
-		s := bufio.NewScanner(m.stdin)
-		for s.Scan() {
-			cid, err := cid.Parse(s.Text())
-			if err != nil {
-				log.Error("error parsing cid", "value", s.Text(), "err", err)
-				return
-			}
-			stream <- cid
-		}
-	}()
-
-	// shutdown gracefully on SIGINT or SIGTERM
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		defer cancel()
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		log.Info("received signal, exiting...")
-	}()
-
-	return indexer.IndexRegisteredWorks(ctx, stream)
-}
-
-func (m *Main) RunERN(args Args) error {
-	switch {
-	case args.Bool("convert"):
-		return m.RunERNConvert(args)
-	case args.Bool("index"):
-		return m.RunERNIndex(args)
-	default:
-		return errors.New("unknown ern command")
-	}
-}
-
-func (m *Main) RunERNConvert(args Args) error {
-	converter := ern.NewConverter(m.store)
-	files := args.List("<files>")
-	for _, file := range files {
-		f, err := os.Open(file)
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		cid, err := converter.ConvertERN(f)
-		if err != nil {
-			return err
-		}
-		fmt.Fprintln(m.stdout, cid.String())
-	}
-	return nil
-}
-
-func (m *Main) RunERNIndex(args Args) error {
-	db, err := sql.Open("sqlite3", args.String("<sqlite3-uri>"))
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	indexer, err := ern.NewIndexer(db, m.store)
-	if err != nil {
-		return err
-	}
-
-	// stream the CIDs from stdin
-	stream := make(chan *cid.Cid)
-	go func() {
-		defer close(stream)
-		s := bufio.NewScanner(m.stdin)
-		for s.Scan() {
-			cid, err := cid.Parse(s.Text())
-			if err != nil {
-				log.Error("error parsing cid", "value", s.Text(), "err", err)
-				return
-			}
-			stream <- cid
-		}
-	}()
-
-	// shutdown gracefully on SIGINT or SIGTERM
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		defer cancel()
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		log.Info("received signal, exiting...")
-	}()
-
-	return indexer.Index(ctx, stream)
 }
 
 func openStore() (*meta.Store, error) {
@@ -482,48 +51,4 @@ func openStore() (*meta.Store, error) {
 		return nil, err
 	}
 	return meta.NewStore(store), nil
-}
-
-type Args map[string]interface{}
-
-func (a Args) String(name string) string {
-	v, ok := a[name]
-	if !ok {
-		panic(fmt.Sprintf("missing arg: %s", name))
-	}
-	if v == nil {
-		return ""
-	}
-	s, ok := v.(string)
-	if !ok {
-		panic(fmt.Sprintf("invalid string arg: %s", name))
-	}
-	return s
-}
-
-func (a Args) List(name string) []string {
-	v, ok := a[name]
-	if !ok {
-		panic(fmt.Sprintf("missing arg: %s", name))
-	}
-	if v == nil {
-		return nil
-	}
-	l, ok := v.([]string)
-	if !ok {
-		panic(fmt.Sprintf("invalid list arg: %s", name))
-	}
-	return l
-}
-
-func (a Args) Bool(name string) bool {
-	v, ok := a[name]
-	if !ok {
-		panic(fmt.Sprintf("missing arg: %s", name))
-	}
-	s, ok := v.(bool)
-	if !ok {
-		panic(fmt.Sprintf("invalid bool arg: %s", name))
-	}
-	return s
 }
